+++ src/main.jsx
@@ -2,15 +2,18 @@ import { StrictMode } from 'react'
 import { createRoot } from 'react-dom/client'
 import { HelmetProvider } from 'react-helmet-async'
 import { ToastProvider } from './app/components/Toast'
+import { AuthProvider } from './app/contexts/AuthContext'
 import './index.css'
 import App from './app/App'
 
 createRoot(document.getElementById('root')).render(
   <StrictMode>
     <HelmetProvider>
-      <ToastProvider>
-        <App />
-      </ToastProvider>
+      <AuthProvider>
+        <ToastProvider>
+          <App />
+        </ToastProvider>
+      </AuthProvider>
     </HelmetProvider>
   </StrictMode>
 )